% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all_compliance.R
\name{all_compliance}
\alias{all_compliance}
\title{Show how many companies send in XML reports yearly}
\usage{
all_compliance(x)
}
\arguments{
\item{x}{A list of data frames from 2010 onwards}
}
\value{
A line graph of annual company filings is generated
}
\description{
Accepts a list of annual company listings, and graphs
the number of companies filing per year.  It does this by counting
the number of rows in each list member, and plotting it against the
year.
}
\details{
This routine is a simple plotting routine that generates an
x-y line graph, where x is the year, starting with 2010, and y is
the number of rows in each data frame that is passed in.
}
\note{
The routine assumes that the first item in the list is the data
for 2010, and each subsequent list item is an additional sequential year.
}
\examples{
all_compliance(list(df1,df2,df3))
}
\author{
Nick Lukianoff
}

