% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backfill_addresses.R
\name{backfill_addresses}
\alias{backfill_addresses}
\title{Backfill addresses}
\usage{
backfill_addresses(x)
}
\arguments{
\item{x}{A data frame in the SEC's format}
}
\value{
A data frame with non-blank address fields
}
\description{
If a business address field is blank, this routine copies
the mailing address into it.  Likewise, if a mailing address field is
blank, it copies the business address into it.
The data is taken from the parsed EDGAR list that the SEC provides.
}
\details{
This routine is provided for continuity of the address data in
10K and 10Q filings.  There are times when only a business address, or
only a mailing address, are provided.  This routine will copy the non-blank
address into the blank address.  The main advantage to this, besides address
data continuity, is in having the business address fields populated.  Some
analytical routines pull their data from the business address, so it's a good
idea to have it fully populated.
The business address fields checked are:  'bas1','bas2','cityba',.stprba','zipba'
The mainling address fields checked are:  'mas1','mas2','cityma',.stprma','zipma'
}
\note{
Since it is highly unlikely that both the maling and business
address fields will be blank at the same time, no check is made to see
if this is the case.
}
\examples{
x <- backfill_addresses(df)
}
\author{
Nick Lukianoff
}

