% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw_state.R
\name{draw_state}
\alias{draw_state}
\title{Map a state's registered companies}
\usage{
draw_state(x, state_input = "", savemap = FALSE)
}
\arguments{
\item{x}{A data frame of addresses in the SEC's format}

\item{state_input}{Uppercase 2-character state code.  If this is not
provided, the function will prompt for a state.}

\item{savemap}{file name of the saved map in .png format.  If this is
not provided, the function will draw the map on the screen,
but not save it to a file.}
}
\value{
A registered company map of the selected state
}
\description{
Displays a map of a state, showing registered companies
clustered by the business zipcode, sized by the amount per zip code,
and color-coded by the company's primary industry.
}
\details{
This function will prompt for a 2-letter uppercase state name (if
there isn't one passed in), then use the data frame's geographic data
(latitude and longitude based on the business zip code) to map the
locations of registered companies in that state.  The map data is shown
by size of data point (based on the amount of companies in a given area),
and color-coded by the industry that it is in.
If no state is entered, Virginia (VA) is assumed.
If no records are found for a given state, an error message is printed.
}
\note{
This function makes use of the ggmap package.
The base SEC address file doesn't contain any geographical data.  This means
that before running this routine, create "Lat" and "Long" columns, and create
a 'companies' column that contains a count of companies per zip code.  The
routine 'add_geodata' can perform this function for you.
}
\examples{
draw_state(df)
draw_state(df,"DE")
draw_state(df,"DE","Delaware Companies")
}
\author{
Nick Lukianoff
}

