% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw_usa.R
\name{draw_usa}
\alias{draw_usa}
\title{Map registered companies in the continental USA}
\usage{
draw_usa(x, savemap = FALSE)
}
\arguments{
\item{x}{A data frame of addresses in the SEC's format}

\item{savemap}{file name of the saved map in .png format.  If this is
not provided, the function will draw the map on the screen,
but not save it to a file.}
}
\value{
A registered company map of the continental US
}
\description{
Displays a map of a the continental US, showing registered
companies clustered by the business zipcode, sized by the amount per zip
code, and color-coded by the company's primary industry.
}
\details{
This function will use the data frame's geographic data
(latitude and longitude based on the business zip code) to map the
locations of registered companies in that state.  The map data is shown
by size of data point (based on the amount of companies in a given area),
and color-coded by the industry that it is in.
}
\note{
This function makes use of the ggmap package.  It also uses the
'Lat' and 'Long' fields to perform its mapping.
The base SEC address file doesn't contain any geographical data.  This means
that before running this routine, create "Lat" and "Long" columns, and create
a 'companies' column that contains a count of companies per zip code.  The
routine 'add_geodata' can perform this function for you.
}
\examples{
draw_usa(df)
draw_usa(df,"Registered Companies")
}
\author{
Nick Lukianoff
}

