% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_states.R
\name{plot_states}
\alias{plot_states}
\title{Interactive bar graph of incorporated companies per state per year}
\usage{
plot_states(x)
}
\arguments{
\item{x}{A list of data frames from 2010 onwards}
}
\value{
An interactive bar graph, selectable by state and year
}
\description{
This function takes a list of address dataframes in the
SEC's format, and graphs each state's incorporated companies per year.
There are options to select any year from the list, or to obtain a
graph of the sum of all of the years.  A dropdown list allows the
subtraction of the states with the most incorporated companies.  This
allows the states with smaller amounts to be more clearly seen.
}
\details{
This routine is a graphing routine that generates a bar
graph, where x is the number of rows per state in each data frame
that is passed in, y is the year, starting with 2010.
A dropdown list is presented listing all of the years represented in
the data frames that are passed in, as well as an option to show all
of the years together.
A second dropdown list is presented with a listing of the largest
state values.  These states can be removed in order to better show
the values of the remaining states.
The states that can be removed are:  DE, NV, MD, CA, NY, FL
}
\note{
The routine assumes that the first item in the list is the data
for 2010, and each subsequent list item is an additional sequential year.
This routine requires the 'manipulate' library.
}
\examples{
plot_states(list(df1,df2,df3))
}
\author{
Nick Lukianoff
}

