% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_usa.R
\name{plot_usa}
\alias{plot_usa}
\title{Interactive map of registered companies in the continental USA}
\usage{
plot_usa(x)
}
\arguments{
\item{x}{A list of data frames from 2010 onwards}
}
\value{
An interactive bar graph, selectable by state and year
}
\description{
This function takes a list of address data froames in
the SEC's format, and maps incorporated companies per year.
There is an option to select any year from the list of passed data frames.
}
\details{
This function will use the data frame's geographic data
(latitude and longitude based on the business zip code) to map the
locations of registered companies in that state.  The map data is shown
by size of data point (based on the amount of companies in a given area),
and color-coded by the industry that it is in.
}
\note{
The routine assumes that the first item in the list is the data
for 2010, and each subsequent list item is an additional sequential year.
This routine removes Alaska and Hawaii companies in order to only display
companies that have a business address in the continental United States.
This routine requires the 'ggmap' library.
This routine requires the 'manipulate' library.
}
\examples{
plot_usa(list(df1,df2,df3))
}
\author{
Nick Lukianoff
}

