% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_nonUSA.R
\name{remove_nonUSA}
\alias{remove_nonUSA}
\title{Remove foreign addresses from address file}
\usage{
remove_nonUSA(x)
}
\arguments{
\item{x}{A data frame in the SEC's format}
}
\value{
A data frame containing only companies registered in a US state
}
\description{
This function strips any data frame row whose
business or mailing addresses don't correspond to a valid US state.
}
\details{
This function will strip away any company that contains any
address that is outside of the United States.  In addition, if the
incorporated state field is empty, it get populated from the business
address, or if the business address is blank, from the mailing address.
}
\note{
This function removes companies from US territories, as well as
foreign companies.
}
\examples{
x <- remove_nonUSA(df)
}
\author{
Nick Lukianoff
}

