% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state_compliance.R
\name{state_compliance}
\alias{state_compliance}
\title{XBRL Reporting Compliance Graph by State}
\usage{
state_compliance(x, state_input = "")
}
\arguments{
\item{x}{A list of data frames from 2010 onwards}

\item{state_input}{Uppercase 2-character state code.  If this is not
provided, the function will prompt for a state.}
}
\value{
A line graph of annual company filings is generated
}
\description{
Accepts a list of annual company listings, and graphs
the number of companies in that state filing per year.  It does this
by counting the number of rows in each list member that correspond to
a given state, and plotting it by year.
}
\details{
This routine is a simple plotting routine that generates an
x-y line graph, where x is the year, starting with 2010, and y is
the number of rows per state in each data frame that is passed in.
If no state is entered, Virginia (VA) is assumed.
If no records are found for a given state, an error message is printed.
}
\note{
The routine assumes that the first item in the list is the data
for 2010, and each subsequent list item is an additional sequential year.
}
\examples{
state_compliance(list(df1,df2,df3))
state_compliance(list(df1,df2,df3), "MD")
}
\author{
Nick Lukianoff
}

