% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state_stat.R
\name{state_stat}
\alias{state_stat}
\title{2 tables of company statistics by state}
\usage{
state_stat(x)
}
\arguments{
\item{x}{A data frame in the SEC's format}
}
\value{
2 tables of company statistics by state
}
\description{
This function lists the number of registered companies
in each of the United States.  The first table lists every state
alphabetically, with a count of incorporated companies.  The second
table lists every state, along with a percentage of companies
incorporated in that state.  The second table is sorted by percentage.
}
\details{
The first thing that this routine does is to clear the screen
by sending a Ctrl-L code to the main device.  It then displays a table,
sorted by state, of companies per state.  A second table follows it,
sorted by percentage of companies in each state.
}
\note{
Keep in mind the Ctrl-L code that precedes the function results.
While this isn't an issue for screen output, it must be dealt with if
you will be storing these 2 tables in a variable.
These tables do not include US Territories, or foreign countries.
}
\examples{
state_stat(df)
x<- state_stat(df)
}
\author{
Nick Lukianoff
}

